move configuration.model.xml test from application to core; remove controlpanel.ui.configuration.test
The controlpanel.ui.configuration only had the XML test, which is now
located in flightrecorder.configuration.test in core. Currently all
tests in both core and application will pass.
