+++ application/tests/pom.xml
@@ -41,7 +41,6 @@
 	<artifactId>missioncontrol.application.tests</artifactId>
 	<packaging>pom</packaging>
 	<modules>
-		<!-- <module>org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.test</module> -->
 		<module>org.openjdk.jmc.flightrecorder.controlpanel.ui.test</module>
 		<module>org.openjdk.jmc.flightrecorder.ext.jfx.test</module>
 		<module>org.openjdk.jmc.flightrecorder.graphview.test</module>
