+++ core/tests/org.openjdk.jmc.common.test/src/main/java/org/openjdk/jmc/common/util/MethodToolkitTest.java
@@ -38,6 +38,7 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+import org.openjdk.jmc.common.util.MethodToolkit;
 
 @SuppressWarnings("nls")
 public class MethodToolkitTest {
