+++ core/tests/org.openjdk.jmc.common.test/src/main/java/org/openjdk/jmc/common/version/JavaVersionTest.java
@@ -30,13 +30,14 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
  * WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.openjdk.jmc.common.version;
+package org.openjdk.jmc.common.test.version;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+import org.openjdk.jmc.common.version.JavaVersion;
 
 public class JavaVersionTest {
