+++ core/tests/org.openjdk.jmc.flightrecorder.configuration.test/src/test/java/org/openjdk/jmc/flightrecorder/configuration/OfflineEventOptionsTest.java
@@ -30,7 +30,7 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
  * WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.openjdk.jmc.flightrecorder.configuration.test;
+package org.openjdk.jmc.flightrecorder.configuration;
 
 import static org.openjdk.jmc.common.unit.UnitLookup.HOUR;
 import static org.openjdk.jmc.common.unit.UnitLookup.MILLISECOND;
@@ -54,7 +54,6 @@ import org.openjdk.jmc.common.unit.LinearUnit;
 import org.openjdk.jmc.common.unit.QuantityConversionException;
 import org.openjdk.jmc.common.unit.QuantityConversionException.Problem;
 import org.openjdk.jmc.common.unit.UnitLookup;
-import org.openjdk.jmc.flightrecorder.configuration.ConfigurationToolkit;
 import org.openjdk.jmc.flightrecorder.configuration.events.EventOptionID;
 import org.openjdk.jmc.flightrecorder.configuration.events.IEventTypeID;
 import org.openjdk.jmc.flightrecorder.configuration.events.SchemaVersion;
