+++ core/tests/org.openjdk.jmc.flightrecorder.configuration.test/src/test/java/org/openjdk/jmc/flightrecorder/configuration/model/xml/XMLTagInstanceTest.java
@@ -30,7 +30,7 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
  * WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml;
+package org.openjdk.jmc.flightrecorder.configuration.model.xml;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -40,10 +40,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.junit.Test;
-import org.openjdk.jmc.flightrecorder.configuration.model.xml.XMLAttribute;
-import org.openjdk.jmc.flightrecorder.configuration.model.xml.XMLNodeType;
-import org.openjdk.jmc.flightrecorder.configuration.model.xml.XMLTag;
-import org.openjdk.jmc.flightrecorder.configuration.model.xml.XMLTagInstance;
 
 @SuppressWarnings("nls")
 public class XMLTagInstanceTest {
